pythonProject7/requirements.txt
pandas
openpyxl
python-docx
requests
matplotlib
numpy
